fix lại lỗi va chạm của quả bóng

+++ src/main/java/Game/MovableObject.java
@@ -1,29 +1,29 @@
 package Game;
 
 public abstract class MovableObject extends GameObject {
-    protected int dx, dy; // tốc độ theo trục X, Y
+    protected double dx, dy; // tốc độ theo trục X, Y
 
     /**
      * Phuong thuc khoi tao
      */
-    public MovableObject(int x, int y, int width, int height, int dx, int dy) {
+    public MovableObject(double x, double y, int width, int height, double dx, double dy) {
         super(x, y, width, height);
         this.dx = dx; this.dy = dy;
     }
 
-    public int getDx() {
+    public double getDx() {
         return dx;
     }
 
-    public void setDx(int dx) {
+    public void setDx(double dx) {
         this.dx = dx;
     }
 
-    public int getDy() {
+    public double getDy() {
         return dy;
     }
 
-    public void setDy(int dy) {
+    public void setDy(double dy) {
         this.dy = dy;
     }
